Moved search bar to independant component
